+++ sentry-android-okhttp/src/main/java/io/sentry/android/okhttp/SentryOkHttpEvent.kt
@@ -4,6 +4,8 @@ import io.sentry.Breadcrumb
 import io.sentry.Hint
 import io.sentry.IHub
 import io.sentry.ISpan
+import io.sentry.SentryDate
+import io.sentry.SentryLevel
 import io.sentry.SpanDataConvention
 import io.sentry.SpanStatus
 import io.sentry.TypeCheckHint
@@ -19,9 +21,12 @@ import okhttp3.Request
 import okhttp3.Response
 import java.util.Locale
 import java.util.concurrent.ConcurrentHashMap
+import java.util.concurrent.RejectedExecutionException
+import java.util.concurrent.atomic.AtomicBoolean
 
 private const val PROTOCOL_KEY = "protocol"
 private const val ERROR_MESSAGE_KEY = "error_message"
+private const val RESPONSE_BODY_TIMEOUT_MILLIS = 500L
 internal const val TRACE_ORIGIN = "auto.http.okhttp"
 
 internal class SentryOkHttpEvent(private val hub: IHub, private val request: Request) {
@@ -29,6 +34,7 @@ internal class SentryOkHttpEvent(private val hub: IHub, private val request: Req
     private val breadcrumb: Breadcrumb
     internal val callRootSpan: ISpan?
     private var response: Response? = null
+    private val isReadingResponseBody = AtomicBoolean(false)
 
     init {
         val urlDetails = UrlUtils.parse(request.url.toString())
@@ -100,13 +106,17 @@ internal class SentryOkHttpEvent(private val hub: IHub, private val request: Req
         // Find the parent of the span being created. E.g. secureConnect is child of connect
         val parentSpan = findParentSpan(event)
         val span = parentSpan?.startChild("http.client.$event") ?: return
+        if (event == RESPONSE_BODY_EVENT) {
+            // We save this event is reading the response body, so that it will not be auto-finished
+            isReadingResponseBody.set(true)
+        }
         span.spanContext.origin = TRACE_ORIGIN
         eventSpans[event] = span
     }
 
     /** Finishes a previously started span, and runs [beforeFinish] on it, on its parent and on the call root span. */
-    fun finishSpan(event: String, beforeFinish: ((span: ISpan) -> Unit)? = null) {
-        val span = eventSpans[event] ?: return
+    fun finishSpan(event: String, beforeFinish: ((span: ISpan) -> Unit)? = null): ISpan? {
+        val span = eventSpans[event] ?: return null
         val parentSpan = findParentSpan(event)
         beforeFinish?.invoke(span)
         if (parentSpan != null && parentSpan != callRootSpan) {
@@ -114,10 +124,11 @@ internal class SentryOkHttpEvent(private val hub: IHub, private val request: Req
         }
         callRootSpan?.let { beforeFinish?.invoke(it) }
         span.finish()
+        return span
     }
 
     /** Finishes the call root span, and runs [beforeFinish] on it. Then a breadcrumb is sent. */
-    fun finishEvent(beforeFinish: ((span: ISpan) -> Unit)? = null) {
+    fun finishEvent(finishDate: SentryDate? = null, beforeFinish: ((span: ISpan) -> Unit)? = null) {
         callRootSpan ?: return
 
         // We forcefully finish all spans, even if they should already have been finished through finishSpan()
@@ -126,7 +137,11 @@ internal class SentryOkHttpEvent(private val hub: IHub, private val request: Req
             it.finish(it.status ?: SpanStatus.INTERNAL_ERROR)
         }
         beforeFinish?.invoke(callRootSpan)
-        callRootSpan.finish()
+        if (finishDate != null) {
+            callRootSpan.finish(callRootSpan.status, finishDate)
+        } else {
+            callRootSpan.finish()
+        }
 
         // We put data in the hint and send a breadcrumb
         val hint = Hint()
@@ -146,4 +161,25 @@ internal class SentryOkHttpEvent(private val hub: IHub, private val request: Req
         RESPONSE_BODY_EVENT -> eventSpans[CONNECTION_EVENT]
         else -> callRootSpan
     } ?: callRootSpan
+
+    fun scheduleFinish(timestamp: SentryDate) {
+        try {
+            hub.options.executorService.schedule({
+                if (!isReadingResponseBody.get() &&
+                    (eventSpans.values.all { it.isFinished } || callRootSpan?.isFinished != true)
+                ) {
+                    finishEvent(timestamp)
+                }
+            }, RESPONSE_BODY_TIMEOUT_MILLIS)
+        } catch (e: RejectedExecutionException) {
+            hub.options
+                .logger
+                .log(
+                    SentryLevel.ERROR,
+                    "Failed to call the executor. OkHttp span will not be finished " +
+                        "automatically. Did you call Sentry.close()?",
+                    e
+                )
+        }
+    }
 }